SSD_WEEK5/backend/init_db.py
from app import create_app
from extensions import db

def init_db():
    app = create_app()
    with app.app_context():
        # Create all tables
        db.create_all()
        print("Database tables created successfully!")

if __name__ == "__main__":
    init_db() 