SSD_WEEK5/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
pg8000==1.29.1
python-dotenv==0.19.0
requests==2.26.0
supabase==2.15.1
asgiref==3.7.2
uvicorn==0.15.0
aiohttp==3.8.1
pandas==1.3.3
numpy==1.21.2
beautifulsoup4==4.9.3
lxml==4.6.3
